requirements.txt
Django==1.9.2
Django-Pushy==0.1.9
Markdown==2.6.5
PyJWT==1.4.0
amqp==1.4.9
anyjson==0.3.3
apns==2.0.1
argparse==1.2.1
billiard==3.3.0.22
celery==3.1.20
django-celery==3.1.17
django-filter==0.12.0
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
kombu==3.0.33
python-gcm==0.4
pytz==2015.7
requests==2.9.1
wsgiref==0.1.2
